+++ deepen/__init__.py
@@ -1,2 +1,3 @@
 from deepen.activation import *
 from deepen.model import *
+from deepen.propagation import *