+++ CHANGELOG.md
@@ -8,10 +8,19 @@ This project uses [Semantic Versioning][sv].
 ### Added
 - Add `learn()` to train the model.
 - Add `predict()` to calculate predictions using the trained model.
+- Create `deepen.propagation` module to provide functions supporting feedforward
+  and backward propagation.
+- Add a Model class with `learn()` and `predict()` methods.
+- Add unit tests for Model.
 
 ### Changed
 - Weights are initialized to have standard deviation of ≈ 1.0 in each layer.
 - Rewrite the computation of the cost function for greater clarity.
+- Move functions supporting feedforward and backward propagation to the new
+  propagation module.
+- Change `predict()` to fetch stored `params` from the model.
+- Change `learn()` to fetch stored `layer_dims` and `learning_rate` from the
+  model.
 
 ## [0.2.0][0.2.0] — 2020-03-30